Allow highcharts to be used in 'style mode'
This adds a way to import the ChartsModule such that it uses the version
of highcharts that can be styled via CSS instead of through highcharts
options.

See http://www.highcharts.com/docs/chart-design-and-style/style-by-css
