+++ karma-test-shim.js
@@ -52,7 +52,8 @@ System.config({
         //'@angular/core/testing': './node_modules/@angular/core/testing',
         '@angular': 'node_modules/@angular',
         // Just stubbing the module with any simple file.
-        'highcharts' : 'dist/stub'
+        'highcharts' : 'dist/stub',
+        'highcharts/js/highcharts' : 'dist/stub'
 
     },
     packages: packages
