+++ src/HighchartsWrapper.ts
@@ -1,4 +1,5 @@
 /// <reference types="highcharts" />
 declare var require: any;
-var Highcharts = require('highcharts');
-export { Highcharts };
+var ClassicModeHighcharts = require('highcharts');
+var StyledModeHighcharts = require('highcharts/js/highcharts');
+export { ClassicModeHighcharts, StyledModeHighcharts };
