+++ src/index.ts
@@ -1,26 +1,12 @@
-import { NgModule } from '@angular/core';
-
+import { ChartModule } from './ChartModule';
 import { ChartComponent } from './ChartComponent';
 import { ChartSeriesComponent } from './ChartSeriesComponent';
 import { ChartPointComponent } from './ChartPointComponent';
 import { ChartXAxisComponent } from './ChartXAxisComponent';
 import { ChartYAxisComponent } from './ChartYAxisComponent';
 
-const CHART_DIRECTIVES: any[] = [
-    ChartComponent,
-    ChartSeriesComponent,
-    ChartPointComponent,
-    ChartXAxisComponent,
-    ChartYAxisComponent
-];
-
-@NgModule({
-  declarations: [CHART_DIRECTIVES],
-  exports: [CHART_DIRECTIVES]
-})
-export class ChartModule { };
-
 export {
+    ChartModule,
     ChartComponent,
     ChartSeriesComponent,
     ChartPointComponent,
